% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmargint-fn.R
\name{deviance.margint}
\alias{deviance.margint}
\title{Deviance for objects of class \code{margint}}
\usage{
\method{deviance}{margint}(object, ...)
}
\arguments{
\item{object}{an object of class \code{margint}, a result of a call to \code{\link{margint.cl}} or \code{\link{margint.rob}}.}

\item{...}{additional other arguments. Currently ignored.}
}
\value{
A real number.
}
\description{
This function returns the deviance of the fitted additive model using one of the three
classical or robust marginal integration estimators, as computed with \code{\link{margint.cl}} or
\code{\link{margint.rob}}.
}
\author{
Alejandra Mercedes Martinez \email{ale_m_martinez@hotmail.com}
}
