% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmargint-fn.R
\name{formula.margint}
\alias{formula.margint}
\title{Additive model formula}
\usage{
\method{formula}{margint}(x, ...)
}
\arguments{
\item{x}{an object of class \code{margint}, a result of a call to \code{\link{margint.cl}} or \code{\link{margint.rob}}.}

\item{...}{additional other arguments. Currently ignored.}
}
\value{
A model formula.
}
\description{
Description of the additive model formula extracted from an object of class \code{margint}.
}
\author{
Alejandra Mercedes Martinez \email{ale_m_martinez@hotmail.com}
}
