% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMI-fn.R
\name{k.epan}
\alias{k.epan}
\title{Epanechnikov kernel}
\usage{
k.epan(x)
kernel2(x)
}
\arguments{
\item{x}{A real number.}
}
\value{
0 if abs(x) > 1 and 0.75 * (1 - x^2) otherwise.
}
\description{
Epanechnikov kernel.
}
\author{
Alejandra Martinez, Matias Salibian-Barrera
}
