% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMI-fn.R
\name{my.norm.2}
\alias{my.norm.2}
\title{Euclidean norm of a vector}
\usage{
my.norm.2(x)
}
\arguments{
\item{x}{A real vector.}
}
\value{
The Euclidean norm of the input vector.
}
\description{
Euclidean norm of a vector.
}
\author{
Alejandra Martinez, Matias Salibian-Barrera
}
