% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmargint-fn.R
\name{print.margint}
\alias{print.margint}
\title{Print a Marginal Integration procedure}
\usage{
\method{print}{margint}(x, ...)
}
\arguments{
\item{x}{an object of class \code{margint}, a result of a call to \code{\link{margint.cl}} or \code{\link{margint.rob}}.}

\item{...}{additional other arguments. Currently ignored.}
}
\value{
A real number.
}
\description{
The default print method for a \code{margint} object.
}
\author{
Alejandra Mercedes Martinez \email{ale_m_martinez@hotmail.com}
}
