% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMI-fn.R
\name{psi.w}
\alias{psi.w}
\title{Tukey bi-square weight function}
\usage{
psi.w(r, k = 4.685)
}
\arguments{
\item{r}{A vector of real numbers.}

\item{k}{Tuning constant.}
}
\value{
A vector of the same length as \code{r}.
}
\description{
This function evaluates the Tukey bi-square weight function defined as Psi(r)/r where Psi is the first derivative of the Tukey loss function. It is used in the re-weighted least squares iterations.
}
\author{
Matias Salibian-Barrera, \email{matias@stat.ubc.ca}, Alejandra Martinez
}
