\name{rmargint-package}
\alias{rmargint-package}
\alias{rmargint}
\docType{package}
\title{
Robust marginal integration estimators for additive models.
}
\description{
Robust marginal integration estimators for additive models.
}
\author{
Alejandra Martinez, Matias Salibian-Barrera

Maintainer: Alejandra Martinez <ale_m_martinez@hotmail.com>
}
\references{
Boente G. and Martinez A. (2017). Marginal integration M-estimators for additive models. TEST, 26, 231-260.}
\keyword{ Marginal integration }