\name{RMI-package}
\alias{RMI}
\docType{package}
\title{
Robust marginal integration procedures for additive models.
}
\description{
Robust marginal integration procedures for additive models.
}
\details{
\tabular{ll}{
Package: \tab RMI\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2018-07-03\cr
License: \tab GPL 3.0\cr
}
}
\author{
Alejandra Martinez, Matias Salibian-Barrera

Maintainer: Alejandra Martinez <alemartinez@unlu.edu.ar>
}
\references{
}
\keyword{ Robust Marginal Integration }
\seealso{
}
\examples{
}
