% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmargint-fn.R
\name{summary.margint}
\alias{summary.margint}
\alias{summary.margint.cl}
\alias{summary.margint.rob}
\title{Summary for additive models fits using a marginal integration procedure}
\usage{
\method{summary}{margint}(object, ...)
}
\arguments{
\item{object}{an object of class \code{margint}, a result of a call to \code{\link{margint.cl}} or \code{\link{margint.rob}}.}

\item{...}{additional other arguments.}
}
\description{
Summary method for class \code{margint}.
}
\details{
This function returns the estimation of the intercept and also the five-number summary and the mean of the residuals for both classical and robust estimators. For the robust estimator it also returns the estimate of the residual standard error.
}
\author{
Alejandra Mercedes Martinez \email{ale_m_martinez@hotmail.com}
}
