% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMI-fn.R
\name{kernel6}
\alias{kernel6}
\title{Order 6 kernel}
\usage{
kernel6(x)
}
\arguments{
\item{x}{A real number.}
}
\value{
0 if abs(x) > 1 and ( 105/256 ) * ( 1 - x^2 ) * ( 5 - 30 * x^2 + 33 * x^4 ) otherwise.
}
\description{
A kernel of order 6.
}
\details{
A kernel L is a kernel of order 6 if it integrates 1, the integrals of u^j L(u) are 0 for 1 <= j < 6 (j integer) and the integral of u^6 L(u) is different from 0.
}
\author{
Alejandra Martinez, Matias Salibian-Barrera
}
