% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMI-fn.R
\name{kernel8}
\alias{kernel8}
\title{Order 8 kernel}
\usage{
kernel8(x)
}
\arguments{
\item{x}{A real number.}
}
\value{
0 if abs(x) > 1 and ( 315/4096 ) * ( 1 - x^2 ) * ( 35 - 385 * x^2 + 1001 * x^4 - 715 * x^6 ) otherwise.
}
\description{
A kernel of order 8.
}
\details{
A kernel L is a kernel of order 8 if it integrates 1, the integrals of u^j L(u) are 0 for 1 <= j < 8 (j integer) and the integral of u^8 L(u) is different from 0.
}
\author{
Alejandra Martinez, Matias Salibian-Barrera
}
