% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMI-fn.R
\name{margint.rob}
\alias{margint.rob}
\title{Robust marginal integration procedures for additive models}
\usage{
margint.rob(Xp, yp, point = NULL, windows, prob = NULL, sigma.hat = NULL,
  win.sigma = NULL, epsilon = 1e-06, type = "0", degree = NULL,
  typePhi = "Huber", k.h = 1.345, k.t = 4.685, max.it = 20,
  qderivate = FALSE, orderkernel = 2, Qmeasure = NULL)
}
\arguments{
\item{Xp}{Matrix of explanatory variables (n by p).}

\item{yp}{Vector of responses (missing values are allowed).}

\item{point}{Matrix of points where predictions will be computed and returned.}

\item{windows}{Vector or a squared matrix of bandwidths for the smoothing estimation procedure.}

\item{prob}{Probabilities of observing each response (n). Defaults to ``NULL''.}

\item{sigma.hat}{Estimate of the residual standard error. If NULL we use the mad of the residuals obtained with local medians.}

\item{win.sigma}{Vector of bandwidths for estimating sigma.hat. If NULL it uses the argument windows if it is a vector or its diagonal if it is a matrix.}

\item{epsilon}{Convergence criterion.}

\item{type}{Three different type of estimators can be selected: type '0' (local constant on all the covariates), type '1' (local linear smoother on all the covariates), type 'alpha' (local polynomial smoother only on the direction of interest).}

\item{degree}{Degree of the local polynomial smoother in the direction of interest when using the estimator of type 'alpha'. Defaults to ``NULL'' for the case when using estimators of type '0' or '1'.}

\item{typePhi}{One of either ``Tukey'' or ``Huber''.}

\item{k.h}{Tuning constant for a Huber-type loss function. Defaults to ``1.345''.}

\item{k.t}{Tuning constant for a Tukey-type loss function. Defaults to ``4.685''.}

\item{max.it}{Maximum number of iterations for the algorithm.}

\item{qderivate}{If TRUE, it calculates g^(q+1)/(q+1)! for each component only for the type 'alpha' method. Defaults to ``FALSE''.}

\item{orderkernel}{Order of the kernel used in the nuisance directions when using the estimator of type 'alpha'. Defaults to ``2''.}

\item{Qmeasure}{A matrix of points where the integration procedure ocurrs. Defaults to ``NULL'' for calcuting the integrals over the sample.}
}
\value{
\item{mu }{Estimate for the intercept.}
\item{g.matrix }{Matrix of estimated additive components (n by p).}
\item{sigma.hat }{Estimate of the residual standard error.}
\item{prediction }{Matrix of estimated additive components for the points listed in the argument point.}
\item{mul }{A vector of size p showing in each component the estimated intercept that considers only that direction of interest when using the type 'alpha' method.}
\item{g.derivative }{Matrix of estimated derivatives of the additive components (only when qderivate is ``TRUE'') (n by p).}
\item{prediction.derivate }{Matrix of estimated derivatives of the additive components for the points listed in the argument point (only when qderivate is ``TRUE'').}
\item{Xp}{Matrix of explanatory variables.}
\item{yp}{Vector of responses.}
}
\description{
Robust marginal integration procedures for additive models.
}
\details{
Three types of robust marginal integration procedures for additive models.
}
\author{
Alejandra Martinez, Matias Salibian-Barrera
}
\keyword{Additive}
\keyword{Models}
