% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMI-fn.R
\name{plot.margint}
\alias{plot.margint}
\title{Diagnostic plots for objects of class \code{margint}}
\usage{
\method{plot}{margint}(object, which = 1:np, ask = FALSE, ...)
}
\arguments{
\item{object}{an object of class \code{margint}, a result of a call to \code{\link{margint.cl}} or \code{\link{margint.rob}}.}

\item{which}{vector of indices of explanatory variables for which partial residuals plots will be generated. Defaults to all available explanatory variables.}

\item{ask}{logical value. If \code{TRUE}, the graphical device will prompt before going to the next page/screen of output.}

\item{...}{additional other arguments.}
}
\description{
Plot method for class \code{margint}.
}
\author{
Alejandra Mercedes Martinez \email{alemartinez@unlu.edu.ar}
}
