% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMI-fn.R
\name{psi.huber}
\alias{psi.huber}
\title{Derivative of Huber's loss function}
\usage{
psi.huber(r, k = 1.345)
}
\arguments{
\item{r}{A real number.}

\item{k}{Tuning constant.}
}
\value{
A real number.
}
\description{
Derivative of Huber's loss function.
}
\details{
The weight function used in the re-weighted least squares iterations.
}
\author{
Alejandra Martinez, Matias Salibian-Barrera
}
