% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMI-fn.R
\name{psi.huber.w}
\alias{psi.huber.w}
\title{Huber weight function}
\usage{
psi.huber.w(r, k = 1.345)
}
\arguments{
\item{r}{A real number.}

\item{k}{Tuning constant.}
}
\value{
A real number.
}
\description{
Huber weight function.
}
\details{
The weight function used in the re-weighted least squares iterations.
}
\author{
Alejandra Martinez, Matias Salibian-Barrera
}
