% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMI-fn.R
\name{psi.tukey}
\alias{psi.tukey}
\title{Derivative of Tukey's bi-square loss function}
\usage{
psi.tukey(r, k = 4.685)
}
\arguments{
\item{r}{A real number.}

\item{k}{Tuning constant.}
}
\value{
A real number.
}
\description{
Derivative of Tukey's bi-square loss function.
}
\author{
Alejandra Martinez, Matias Salibian-Barrera
}
