% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMI-fn.R
\name{psi.w}
\alias{psi.w}
\title{Tukey bi-square weight function}
\usage{
psi.w(r, k = 4.685)
}
\arguments{
\item{r}{A real number.}

\item{k}{Tuning constant.}
}
\value{
A real number.
}
\description{
Tukey bi-square weight function.
}
\details{
he weight function used in the re-weighted least squares iterations.
}
\author{
Alejandra Martinez, Matias Salibian-Barrera
}
