% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMI-fn.R
\name{residuals.margint}
\alias{residuals.margint}
\title{Residuals of a marginal integration fit in an additive model}
\usage{
\method{residuals}{margint}(object, ...)
}
\arguments{
\item{object}{an object of class \code{margint}, a result of a call to \code{\link{margint.cl}} or \code{\link{margint.rob}}.}

\item{...}{additional other arguments.}
}
\value{
A vector of residuals.
}
\description{
This function returns the residuals of the fitted additive model using one of the three
classical or robust marginal integration estimators, as computed with \code{\link{margint.cl}} or
\code{\link{margint.rob}}.
}
\author{
Alejandra Mercedes Martinez \email{alemartinez@unlu.edu.ar}
}
