\name{rmargint-package}
\alias{rmargint-package}
\alias{rmargint}
\docType{package}
\title{
Robust marginal integration estimators for additive models.
}
\description{
Robust marginal integration estimators for additive models.
}
\details{
\tabular{ll}{
Package: \tab rmargint\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2019-10-15\cr
License: \tab GPL 3.0\cr
}
}
\author{
Alejandra Martinez, Matias Salibian-Barrera

Maintainer: Alejandra Martinez <ale_m_martinez@hotmail.com>
}
\references{
Boente G. and Martinez A. (2017). Marginal integration M-estimators for additive models. TEST, 26, 231-260.}
\keyword{ Marginal integration }