\name{RMI-package}
\alias{RMI-package}
\alias{RMI}
\docType{package}
\title{
Robust marginal integration procedures for additive models
}
\description{
Robust marginal integration procedures for additive models
}
\details{
\tabular{ll}{
Package: \tab RMI\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2019-06-10\cr
License: \tab GPL 3.0\cr
}
}
\author{
Alejandra Martinez, Matias Salibian-Barrera

Maintainer: Alejandra Martinez <ale_m_martinez@hotmail.com>
}
\references{
Boente G. and Martínez A. (2017). Marginal integration M−estimators for additive models. TEST, 26, 231-260.}
\keyword{ Marginal integration }